pyproject.toml
[tool.poetry]
name = "eventpulse"
version = "0.1.0"
description = ""
authors = ["vantutran2k1 <vantu.tran2k1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
faker = "^37.6.0"
confluent-kafka = "^2.11.1"
requests = "^2.32.5"
python-dotenv = "^1.1.1"
orjson = "^3.11.3"
avro = "^1.12.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
